p_graphic: single-precision t0/t1 color interp in drawScreenFade (80.2%) (#16848)

+++ src/p_graphic.cpp
@@ -414,14 +414,14 @@ void CGraphicPcs::drawScreenFade()
                                                GX_LINEAR, GX_TF_RGBA8, 0);
                     GXLoadTexObj(&backTexObj, GX_TEXMAP0);
 
-                    const float t0 = (float)((double)row * kGraphicHalf);
+                    const float t0 = (float)row * kGraphicHalf;
                     CColor topColor;
                     topColor.color.r = (u8)(t0 * ((float)baseColor2.r - (float)baseColor.r) + (float)baseColor.r);
                     topColor.color.g = (u8)(t0 * ((float)baseColor2.g - (float)baseColor.g) + (float)baseColor.g);
                     topColor.color.b = (u8)(t0 * ((float)baseColor2.b - (float)baseColor.b) + (float)baseColor.b);
                     topColor.color.a = 0xFF;
 
-                    const float t1 = (float)((double)(row + 1) * kGraphicHalf);
+                    const float t1 = (float)(row + 1) * kGraphicHalf;
                     CColor bottomColor;
                     bottomColor.color.r = (u8)(t1 * ((float)baseColor2.r - (float)baseColor.r) + (float)baseColor.r);
                     bottomColor.color.g = (u8)(t1 * ((float)baseColor2.g - (float)baseColor.g) + (float)baseColor.g);
